auth-service/requirements.txt
flask==3.0.0
PyJWT==2.8.0
mysql-connector-python==8.3.0
